Simplify head behaviour (#828)
