+++ module/behaviour/skills/HeadBehaviourSoccer/src/HeadBehaviourSoccer.hpp
@@ -20,134 +20,40 @@
 #ifndef MODULES_BEHAVIOUR_REFLEX_HEADBEHAVIOURSOCCER_HPP
 #define MODULES_BEHAVIOUR_REFLEX_HEADBEHAVIOURSOCCER_HPP
 
+#include <Eigen/Core>
 #include <nuclear>
-#include <set>
-
-#include "Searcher.hpp"
-
-#include "message/behaviour/SoccerObjectPriority.hpp"
-#include "message/input/Image.hpp"
-#include "message/input/Sensors.hpp"
-#include "message/localisation/Ball.hpp"
-#include "message/motion/HeadCommand.hpp"
-#include "message/motion/KinematicsModel.hpp"
-#include "message/vision/Ball.hpp"
-#include "message/vision/Goal.hpp"
-
-#include "utility/math/geometry/Quad.hpp"
 
 namespace module::behaviour::skills {
 
     /**
      * Executes a HeadBehaviourSoccer action.
      *
-     * @author Jade Fountain
+     * @author Thomas O'Brien
      */
     class HeadBehaviourSoccer : public NUClear::Reactor {
     public:
     private:
-        enum SearchState { FIXATION = 0, WAIT = 1, SEARCH = 2 };
-        SearchState state = SearchState::SEARCH;
+        // Time before starting to search for ball after its lost
+        float search_timeout_ms = 0.0f;
+        // Time lingering at each position in lost ballsearch
+        float fixation_time_ms = 0.0f;
+        // Index in the list of search positions
+        long unsigned int searchIdx = 0;
 
-        message::vision::Balls getFixationObjects(const std::shared_ptr<const message::vision::Balls>& vballs,
-                                                  bool& search);
-        message::vision::Goals getFixationObjects(const std::shared_ptr<const message::vision::Goals>& vgoals,
-                                                  bool& search);
-
-
-        /*! @brief Updates the search plan when something has changed
-         */
-        void updateHeadPlan(const message::motion::KinematicsModel& kinematicsModel,
-                            const message::vision::Balls& fixationObjects,
-                            const bool& search,
-                            const message::input::Sensors& sensors,
-                            const Eigen::Matrix3d& headToIMUSpace,
-                            const message::input::Image::Lens& lens);
-        void updateHeadPlan(const message::motion::KinematicsModel& kinematicsModel,
-                            const message::vision::Goals& fixationObjects,
-                            const bool& search,
-                            const message::input::Sensors& sensors,
-                            const Eigen::Matrix3d& headToIMUSpace,
-                            const message::input::Image::Lens& lens);
-
-        /*! @brief Converts from camera space direction to IMU space direction
-         */
-        static Eigen::Vector2d getIMUSpaceDirection(const Eigen::Vector2d& screenAngles,
-                                                    const Eigen::Matrix3d& headToIMUSpace);
-
-        /*! @brief Gets points which allow for simultaneous search and viewing of key objects
-         */
-        std::vector<Eigen::Vector2d> getSearchPoints(const message::motion::KinematicsModel& kinematicsModel,
-                                                     const message::vision::Balls& fixationObjects,
-                                                     const message::behaviour::SoccerObjectPriority::SearchType& sType,
-                                                     const message::input::Sensors& sensors,
-                                                     const message::input::Image::Lens& lens);
-        std::vector<Eigen::Vector2d> getSearchPoints(const message::motion::KinematicsModel& kinematicsModel,
-                                                     const message::vision::Goals& fixationObjects,
-                                                     const message::behaviour::SoccerObjectPriority::SearchType& sType,
-                                                     const message::input::Sensors& sensors,
-                                                     const message::input::Image::Lens& lens);
-
-        /*! @brief Combines a collection of vision objects. The screen resulting screen angular region is the
-         * bounding box of the objects
-         */
-        message::vision::Ball combineVisionObjects(const message::vision::Balls& obs);
-        message::vision::Goal combineVisionObjects(const message::vision::Goals& obs);
-
-        /*! @brief Gets a bounding box in screen angular space of a set of vision objects
-         */
-        static utility::math::geometry::Quad<double, 2, 1> getScreenAngularBoundingBox(
-            const message::vision::Balls& obs);
-        static utility::math::geometry::Quad<double, 2, 1> getScreenAngularBoundingBox(
-            const message::vision::Goals& obs);
-
-        bool orientationHasChanged(const message::input::Sensors& sensors);
-
-
-        // CONFIG - loaded elsewhere
-        float max_yaw   = 0.0f;
-        float min_yaw   = 0.0f;
-        float max_pitch = 0.0f;
-        float min_pitch = 0.0f;
-
-
-        float replan_angle_threshold = 0.0f;
-        Eigen::Matrix3d Rtw{};
-
-        // CONFIG from HeadBehaviourSoccer.yaml
-        float pitch_plan_threshold                = 0.0f;
-        float pitch_plan_value                    = 20.0f;
-        double fractional_view_padding            = 0.0;
-        float search_timeout_ms                   = 0.0f;
-        float fractional_angular_update_threshold = 0.0f;
-
-        bool oscillate_search = false;
-
-        bool locBallReceived = false;
-        message::localisation::Ball lastLocBall{};
-
-        std::map<message::behaviour::SoccerObjectPriority::SearchType, std::vector<Eigen::Vector2d>> searches{};
-
-        // State variables
-        Searcher<Eigen::Vector2d> headSearcher{};
-
-        int ballPriority = 0;
-        int goalPriority = 0;
-        message::behaviour::SoccerObjectPriority::SearchType searchType =
-            message::behaviour::SoccerObjectPriority::SearchType::LOST;
-
-        NUClear::clock::time_point lastPlanUpdate{};
-        NUClear::clock::time_point timeLastObjectSeen{};
-
-        Eigen::Vector2d lastCentroid = Eigen::Vector2d::Zero();
-
-        bool lostAndSearching = false;
-        bool lostLastTime     = false;
-
+        // Flag for if the robot is currently getting up
         bool isGettingUp = false;
 
-        int lastBallPriority = 0;
-        int lastGoalPriority = 0;
+        // Time between last search position transition
+        NUClear::clock::time_point searchLastMoved = NUClear::clock::now();
+
+        // List of positions for search
+        std::vector<Eigen::Vector2d> search_positions;
+
+        // Time since last ball seen
+        NUClear::clock::time_point ballLastMeasured = NUClear::clock::now();
+
+        // Position of last seen ball
+        Eigen::Vector3d rBCc = Eigen::Vector3d::Zero();
 
     public:
         explicit HeadBehaviourSoccer(std::unique_ptr<NUClear::Environment> environment);
