+++ module/behaviour/skills/HeadBehaviourSoccerOld/CMakeLists.txt
@@ -1,0 +1,2 @@
+# Build our NUClear module
+nuclear_module()
