+++ module/behaviour/skills/HeadBehaviourSoccerOld/data/config/nugus1/HeadBehaviourSoccerOld.yaml
