+++ module/motion/HeadController/src/HeadController.hpp
@@ -33,22 +33,18 @@ namespace module::motion {
     class HeadController : public NUClear::Reactor {
     private:
         const size_t id;
-        double min_yaw, max_yaw, min_pitch, max_pitch, head_motor_gain, head_motor_torque, p_gain;
+        float min_yaw, max_yaw, min_pitch, max_pitch, head_motor_gain, head_motor_torque, smoothing_factor;
         ReactionHandle updateHandle;
         // Debug var:
         NUClear::clock::time_point lastTime;
 
     public:
-        static constexpr const char* CONFIGURATION_PATH = "HeadController.yaml";
-        static constexpr const char* CONFIGURATION_MSSG = "Head Controller - Configure";
-        static constexpr const char* ONTRIGGER_HEAD_CMD = "Head Controller - Register Head Command";
-        static constexpr const char* ONTRIGGER_HEAD_POS = "Head Controller - Update Head Position";
-
         explicit HeadController(std::unique_ptr<NUClear::Environment> environment);
 
         Eigen::Vector2f currentAngles;
         Eigen::Vector2f goalAngles;
         bool goalRobotSpace = true;
+        bool smooth         = true;
     };
 }  // namespace module::motion
