+++ module/motion/KinematicsConfiguration/data/config/KinematicsConfiguration.yaml
@@ -26,7 +26,7 @@ head:
   # head movement limits
   limits: # pitch limit is a function of yaw ... need to calculate this properly
     yaw: [-pi/4, pi/4] # [min_yaw, max_yaw]
-    pitch: [0, pi/2] # [min_pitch, max_pitch]
+    pitch: [-pi/2, pi/8] # [min_pitch, max_pitch]
 # rough measurements
 arm:
   distance_between_shoulders: 0.17 # distance between shoulders
